Factory is now responsible for workspace knowledge
