+++ .eslintrc
@@ -22,6 +22,12 @@
         }],
         "@typescript-eslint/no-require-imports": 2,
         "no-duplicate-imports": ["error"],
+        "sort-imports": ["error", {
+            "ignoreCase": true,
+            "ignoreDeclarationSort": true,
+            "ignoreMemberSort": false,
+            "memberSyntaxSortOrder": ["none", "all", "multiple", "single"]
+        }],
         "comma-dangle": [
             "error",
             {
@@ -54,6 +60,7 @@
         "keyword-spacing": "error",
         "no-sync": "error",
         "no-lonely-if": "error",
-        "arrow-spacing": "error"
+        "arrow-spacing": "error",
+        "eol-last": ["error", "always"]
     }
 }
