--- src/util/git.api.interface.ts
@@ -1,7 +1,0 @@
-export interface VscodeGitExtension {
-    git: GitPath;
-}
-
-export interface GitPath {
-    path: string;
-}
